Bug #909533: Swift uploads through Glance using ridiculously small chunks
Fix the parsing of the swift_store_large_object_size and
swift_store_large_object_chunk_size options.  These are specified in MB in
the config file, but need to be in bytes internally, because they are
compared against the image_size parameter given to add().

Change the streaming to read from the wire 64K at a time, and accumulate that
in a file on disk.  The avoids the previous behavior where a 200 MB buffer
would be allocated per stream, because that's how much we were asking
image_file.read() for.

Update the unit tests to match.  This includes an additional check that
put_object is called the correct number of times.  The unit tests missed this
problem because there was no check that a small object is only uploaded in
one chunk, and despite the comment to the contrary there was no check that
the right number of chunks were written in the large object case either.